requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-multipart
pdfplumber
python-docx
beautifulsoup4
requests
